Move Builder as non inner class in async-processor
